[challange-3] - Splitting Personal/Business expenses into different section on UI
